requirements.txt
pyTelegramBotAPI==4.24.0
python-dotenv==1.0.1
telegramify_markdown==0.1.17
aiohttp==3.10.10
openai==1.61.1
